requirements.txt
halftone==0.0.1
Pillow==10.1.0